refactor: extract partition definition preprocessing
issue: #39
